+++ docker-compose.yml
@@ -79,6 +79,29 @@ services:
               capabilities: [ gpu ]
     command: [ "uvicorn", "agent_zephyr:app", "--host", "0.0.0.0", "--port", "8000" ]
 
+  bloom_api:
+    build:
+      context: models/bloom
+    container_name: bloom_api
+    runtime: nvidia
+    ports:
+      - "8005:8000"
+    networks:
+      - llm-network
+    environment:
+      NVIDIA_VISIBLE_DEVICES: 1
+      NVIDIA_DRIVER_CAPABILITIES: compute,utility
+      LD_LIBRARY_PATH: /usr/lib/x86_64-linux-gnu:/usr/local/cuda/lib64
+    restart: always
+    deploy:
+      resources:
+        reservations:
+          devices:
+            - driver: nvidia
+              count: 1
+              capabilities: [ gpu ]
+    command: [ "uvicorn", "agent_bloom:app", "--host", "0.0.0.0", "--port", "8000" ]
+
   router_api:
     build:
       context: ./router
@@ -89,6 +112,7 @@ services:
       - distillbert_api
       - llama_api
       - zephyr_api
+      - bloom_api
     ports:
       - "8080:8000"
     networks:
