+++ models/bloom/Dockerfile
@@ -22,4 +22,4 @@ ENV PATH="/app/venv/bin:$PATH"
 
 EXPOSE 8000
 
-CMD ["uvicorn", "agent_llama:app", "--host", "0.0.0.0", "--port", "8000"]
+CMD ["uvicorn", "agent_bloom:app", "--host", "0.0.0.0", "--port", "8000"]
