+++ models/bloom/requirements.txt
@@ -1,6 +1,6 @@
+torch
 fastapi
 uvicorn
 transformers
 accelerate
-protobuf
-torch
+bitsandbytes
