+++ router/router.py
@@ -11,6 +11,7 @@ MODELS = {
     "gbert": "http://gbert_api:8000/process/",
     "llama": "http://llama_api:8000/process/",
     "zephyr": "http://zephyr_api:8000/process/",
+    "bloom": "http://bloom_api:8000/process/",
 }
 
 class TextRequest(BaseModel):
@@ -29,11 +30,12 @@ async def process_text(request: TextRequest):
     text = request.text
 
     try:
-        distillbert_result, gbert_result, llama_result, zephyr_result = await asyncio.gather(
+        distillbert_result, gbert_result, llama_result, zephyr_result, bloom_result = await asyncio.gather(
             get_model_response("distillbert", text),
             get_model_response("gbert", text),
             get_model_response("llama", text),
             get_model_response("zephyr", text),
+            get_model_response("bloom", text),
         )
 
         prompt = f"Verbessere den folgenden Text auf Deutsch unter Berücksichtigung der sprachlichen Qualität ({distillbert_result.get('sprachliche_qualitaet')}) und der formalen Vorgaben ({gbert_result.get('formale_vorgaben')}): {text}"
@@ -49,6 +51,7 @@ async def process_text(request: TextRequest):
             "formale_vorgaben": gbert_result.get("formale_vorgaben"),
             "gliederung": llama_result.get("gliederung"),
             "citation": zephyr_result.get("citation"),
+            "corrected_grammar": bloom_result.get("checked_text"),
         }
 
         gewichtetes_ergebnis = 0
